Rename all instances of teamXXX to team37
