+++ pom.xml
@@ -10,7 +10,7 @@
     If you're team 542, use team542 . You got the idea!
   -->
   
-  <artifactId>teamXXX</artifactId>
+  <artifactId>team37</artifactId>
 
   <!-- Don't edit the POM file below this line. Ask instructor if you need to (you should not need it) -->
