+++ src/main/java/ca/mcmaster/se2aa4/island/team37/actions/ActionManager.java
@@ -1,8 +1,8 @@
-package ca.mcmaster.se2aa4.island.teamXXX.actions;
+package ca.mcmaster.se2aa4.island.team37.actions;
 
 import org.json.JSONObject;
 
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Direction;
+import ca.mcmaster.se2aa4.island.team37.drone.Direction;
 
 /**
  * The ActionManager class is responsible for creating JSON strings that
