+++ src/test/java/ca/mcmaster/se2aa4/island/team37/ExplorerFactory.java
@@ -2,6 +2,8 @@ package ca.mcmaster.se2aa4.island.teamXXX;
 
 import org.json.JSONObject;
 
+import ca.mcmaster.se2aa4.island.team37.Explorer;
+
 public class ExplorerFactory {
 
     public Explorer createExplorer(String parameters) {
